Lots of changes in this here first-ish commit.
First of all, the ScenarioModule can now load and save its data properly. There is a skeleton data system set up for the categories, how to handle parents, and how to assign vessels to parents.

Next step is to start working on a GUI so one can actually add categories properly. The main idea is that vessels are permitted only one category.