Toying around with targeting vessels. No luck..
Seems targeting vessels is easy in flight, but in the tracking center, it is a whole other story. Tried a couple of ways now, but none of them seem to work out. The PlanetariumCamera target list in the tracking station seems to be empty, which is odd. Tried adding a target, still no luck. Ah well, committing this for now.