Added basic switching between vessels
By using some snippets from HaystackContinued (https://github.com/qberticus/HaystackContinued) I have managed to get switching to work properly. I also modified the code for showing the "Control" button. Will probably replace this with another button later, but for now it works.

This also lets me focus the map camera while in the tracking station. Unfortunately, it does not seem to work with the in flight map. More experimentation needed for that.